faster/services/music/requirements.txt
fastapi
sqlalchemy
pillow
requests
alembic
pydantic
mutagen
psycopg2-binary
asyncpg
pytest
pydantic-settings
uvicorn
pytest-asyncio
pytest-tornasync
python-dotenv
